ext/intl: use pkg-config to detect icu
icu-config is deprecated and not available in some distros anymore. This
patch tries to setup icu with pkg-config first. It's a smaller patch
than the original 20fa2e7b58c519cc148d9658456b695884b1ecf4 by Hugh
McMaster. This one does not depend on the PKG_CHECK_MODULES macro so it
is more suitable to patch PHP 7.0 and older.